all title widgets goes to TitleWidget.hpp/TitleWidget.cpp
