--- Client/Client.Qt/Widgets/BioButton.hpp
@@ -1,24 +1,0 @@
-#pragma once
-
-#include "Style/StyleBasic.hpp"
-#include "Widgets/CaptionButton.hpp"
-#include "Widgets/PopupWidget.hpp"
-
-/**
- * @class BioButton
- * @brief Class for Profile menu widget
- */
-class BioButton : public CaptionButton
-{
-    Q_OBJECT
-public:
-    /**
-     * @brief Constructor of bio button
-     * @param parent Parent of widget
-     * @param inCaption Must be set true for now
-     */
-    explicit BioButton(QWidget* parent = nullptr, bool inCaption = true);
-
-private:
-    std::unique_ptr<PopupWidget> _popup;
-};
