+++ Client/Client.Qt/login.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
 #include <QWidget>
+#include <memory>
+
+#include "Widgets/LogoWidget.hpp"
 
 class FlatInput;
 class FlatButton;
-class LogoWidget;
 class LinkButton;
 
 /**
@@ -29,8 +31,8 @@ protected:
     void resizeEvent(QResizeEvent* event) override;
 
 private:
-    std::unique_ptr<FlatInput> _usernameInput;
-    std::unique_ptr<FlatInput> _passwordInput;
+    std::unique_ptr<FlatInput>  _usernameInput;
+    std::unique_ptr<FlatInput>  _passwordInput;
     std::unique_ptr<FlatButton> _signInButton;
     std::unique_ptr<FlatButton> _registrationButton;
     std::unique_ptr<LinkButton> _forgotPasswordLink;
